Fixed encoding issues when handling non ascii files
Added MassEdit.encoding member and --encoding option to specify the
encoding of the input and output files (defaults to UTF-8). MassEdit
should honor the encoding when reading and writing files.

When writing the diff to an output, MassEdit now rounds trip to the
encoding specified by the output to avoid UnicodeError exceptions.

Fixed other minor issues.